skills_list_be/requirements.txt
Django==5.1.5
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
PyJWT==2.10.1
pytz==2025.1
six==1.17.0